bfloat cast issue

+++ src/recsys/model.py
@@ -242,7 +242,7 @@ class MultitaskRecommender(LightningModule):
 
         loss = self.criterion(scores, labels)
         
-        accuracy = self.accuracy(scores, labels)
+        accuracy = self.accuracy(scores.float(), labels.float())
         self.log("val_accuracy", accuracy, prog_bar=True) 
         self.log("val_loss", loss, prog_bar=True)
         self.predictions.append(scores.detach().cpu().flatten().float().numpy())
